Restructure Marconi-Mamba modules
- Renamed the module names to match marconi-chain-index
- Renamed http handlerss and route to match marconi-chain-index
- Moved marconi-chain-index property generators to marconi-chain-index-test-lib
- Add CLI golden tests

[PLT-1360]

 Re-org moconi/marconi-mamba test folders

Changed test folder structure to match those of AddressDatum Spec standard

[PLT-1360]
