+++ marconi-chain-index/test/Spec/Marconi/ChainIndex/Indexers/EpochStakepoolSize.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE TupleSections      #-}
 {-# LANGUAGE TypeApplications   #-}
 
-module EpochStakepoolSize where
+module Spec.Marconi.ChainIndex.Indexers.EpochStakepoolSize (tests) where
 
 import Control.Concurrent qualified as IO
 import Control.Concurrent.Async qualified as IO
