+++ marconi-mamba/test/Spec/Golden/Cli/marconi___addresses_to_index.help
@@ -1,0 +1,6 @@
+The option `--addresses-to-index` expects an argument.
+
+Usage: marconi-mamba --socket-path FILE-PATH --utxo-db FILE-PATH 
+                     [--http-port HTTP-PORT] 
+                     (--mainnet | --testnet-magic NATURAL)
+                     (--addresses-to-index BECH32-ADDRESS )
