+++ marconi-mamba/test/Spec/Golden/Cli/marconi___help.help
@@ -1,0 +1,6 @@
+Invalid option `--help'
+
+Usage: marconi-mamba --socket-path FILE-PATH --utxo-db FILE-PATH 
+                     [--http-port HTTP-PORT] 
+                     (--mainnet | --testnet-magic NATURAL)
+                     (--addresses-to-index BECH32-ADDRESS )
